+++ changelog.md
@@ -2,6 +2,7 @@ v0.0.11-alpha
 -------------
 * Update research table project background
 * Made Flint and Steel set Treefolk on fire
+* Add requirement of Sea Dust to Cryotreatment research
 
 v0.0.10-alpha
 -------------