+++ changelog.md
@@ -11,6 +11,7 @@ v0.0.11-alpha
 * Reduced effectiveness of Quicken spell mod to 0.15sec per level, from 0.5sec per level
 * Increase damage and healing of all spell types
 * Double duration of all damage spell secondary effects
+* Rebalanced mana costs of most spells
 
 v0.0.10-alpha
 -------------